pyproject.toml
[project]
name = "airflow-github-issues"
version = "0.1.0"
description = "Fetches issues from the Apache Airflow GitHub repository and stores them in Elasticsearch."
dependencies = [
    "requests",
    "elasticsearch",
    "python-dotenv>=1.2.2",
    "tqdm>=4.67.3",
]
requires-python = ">=3.11"
